+++ src/Pegasus/Common/CIMInstanceRep.h
@@ -65,7 +65,7 @@ public:
 
     Boolean existsQualifier(const String& name) const
     {
-	return ((_qualifiers.find(name) != -1) ? true : false);
+	return ((_qualifiers.find(name) != Uint32(-1)) ? true : false);
     }
